chore: break out trade input components for re-use
